+++ api/device-tracker.proto
@@ -2,7 +2,7 @@ syntax = "proto3";
 
 package device_tracker;
 
-option go_package = "github.com/dak151449/IOT_Device_Tracker_Service/pkg/device_tracker;api";
+option go_package = "github.com/dak151449/IOT_Device_Tracker_Service/pkg/device-tracker;api";
 
 service DeviceTrackerService {
 	rpc ToUpper(ToUpperRequest) returns (ToUpperResponse) {}
