+++ internal/app/dtservice/to_upper.go
@@ -1,4 +1,4 @@
-package dt_service
+package dtservice
 
 import (
 	"context"
